==> 02_extracting_terminology.ipynb <==
# Extracting terminology (key words and phrases)

In order to build a glossary, we need to know what words and phrases qualify as terms. For that I will be processing the source (EN) part of the parallel corpus, and later look for matches in the target (RU) part.

%run utility_file     # handles module imports and loading .csv files
from utility_file import Preprocess     # custom class for preprocessing text

path = 'pi2.csv'
source_lang = 'English'
target_lang = 'Russian'

source_list, target_list = load_separate_corpora_from_csv(path, source_lang, target_lang)

exceptions = [
    'flowerbed',
    'building',
    ]

clean_en_list = [Preprocess(sentence).preprocess_no_lemmatization(exceptions) \
                 for sentence in source_list if len(sentence) <= 80]
# only looking at strings < 80 symbols as longer strings are less likely to contain terms and will clutter the corpus
clean_en_str = ' '.join(clean_en_list)

## Extracting unigrams (one-word terms)

The first step is simple - I will look at the entire corpus and extract all the verbs and nouns using [spaCy morphology analyzer](https://spacy.io/api/morphology).

# creating sets of all nouns and verbs using spaCy

doc = nlp(clean_en_str)
nouns = set()
verbs = set()
ads = set()
for token in doc:
    if token.pos_ == "NOUN":
        nouns.add(token.lemma_)
    elif token.pos_ == "VERB":
        if token.text != "\'ve" and token.text != "s" and token.text != "\'m" and token.text != "\'re":
            if len(token.text) > 2:
                verbs.add(token.lemma_)
all_pos = nouns.union(verbs)

Then I'm extracting potential one-word terms using the [tf-idf](https://en.wikipedia.org/wiki/Tf%E2%80%93idf) algorithm provided by [sklearn](https://scikit-learn.org/stable/modules/generated/sklearn.feature_extraction.text.TfidfVectorizer.html).

# Extracting candidate words using TfidfVectorizer

stop_words = "english"
count = TfidfVectorizer(min_df = 5, stop_words=stop_words).fit(clean_en_list)
bag_of_words = count.transform(clean_en_list)
sum_words = bag_of_words.sum(axis=0)
words_freq = [(word, sum_words[0, idx]) for word, idx in count.vocabulary_.items()]
words_freq =sorted(words_freq, key = lambda x: x[1], reverse=True)
tfidf_candidates = []
for item in words_freq[:700]:      # only looking at 700 most common words
    tfidf_candidates.append(item[0])

As I am only interested in nouns and verbs at this stage, I'm creating the final list of unigram terms by intersecting the set of nouns and terms created previously and the list of tf-idf candidates.

# Intersecting with the list of nouns and verbs
candidates = [x for x in tfidf_candidates if x in all_pos]

# taking a look
print(candidates[:100])

## Extracting n-grams (key phrases)

The idea here is the same as with unigrams - first, extracting potential terms with a linguistic tool (spaCy [noun chunker](https://spacy.io/usage/linguistic-features#noun-chunks) in this case).

# extracting n-grams using spacy chunker

noun_chunks = set(chunk.text.strip().lower() for chunk in doc.noun_chunks)

Then extracting n-gram key phrases. In this case, I will be using sklearn's [CountVectorizer](https://scikit-learn.org/stable/modules/generated/sklearn.feature_extraction.text.CountVectorizer.html) rather than TfidfVectorizer, for this algorithm seems to be providing better results.

# Extracting candidate n-grams with CountVectorizer and intersect them with the noun chunks

n_gram_range = (2, 3)
count = CountVectorizer(ngram_range=n_gram_range, stop_words=stop_words).fit([clean_en_str])
ngrams = count.get_feature_names()

As previously, creating the final list of n-gram terms by intersecting the two lists. 

ngram_candidates = list(filter(lambda candidate: candidate in noun_chunks, ngrams))

print(ngram_candidates[:50])

## Combining + Pickling

keywords = set(candidates + ngram_candidates)
print(len(keywords), 'is the total number of extracted terms')

# pickling the keyword set

import pickle
with open('keywords.pkl', 'wb') as f:
       pickle.dump(set(keywords), f)
